test: complete serialize bytes tests
The remaining problem is the difference in the way the official class index is assigned. Several possibilities have emerged from previous research.

- The index is incremented each time the class is deserialized, including class inheritance.
- Classes including pointers are deserialized later.

However, this is quite difficult to achieve at this time.
Closures would be needed for later deserialization, but using closures in Rust often causes problems that do not have a lifetime.